requirements.txt
python-telegram-bot
telethon
yt-dlp
python-dotenv
